Extract TodoItemTaskInput.tsx
set focus on input
